More specific

+++ library/c_library/david_sum.c
@@ -1,3 +1,5 @@
-int david_sum(int a, int b) {
+#include <stdint.h>
+
+uint8_t david_sum(uint8_t a, uint8_t b) {
 	return a + b;
 }
