+++ library/c_library/david_sum.h
@@ -1,1 +1,1 @@
-int david_sum(int, int);
+uint8_t david_sum(uint8_t, uint8_t);
